fix: Hot 게시글 삭제 batch -> scheduler로 수정, 삭제 쿼리 수정
